fix: prefer fresh access after checkout ready
